bug fix with typedef of line width in giza fill (thanks to W. Dehnen)
git-svn-id: https://svn.code.sf.net/p/giza/code@458 952f94de-64c9-43ce-8081-c12abd9885bf

+++ src/giza-fill.c
@@ -99,7 +99,7 @@ _giza_fill (void)
   double ri,gi,bi,alphai;
   int ci;
   int hatch_size;
-  int lw;
+  double lw;
   double xoffset,dx,dy,dr;
   double pi = 3.1415926536;
   double xmin,xmax,ymin,ymax,x0,y0,x,y,angle;
